--- backend/routes.js
@@ -1,1 +1,0 @@
-//make routes (for tmr)
